pyproject.toml
[tool.poetry]
name = "tg-avito-bot"
version = "0.1.0"
description = ""
authors = ["Ilmir Nigmatullin <nigmatullin.mir@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.13.1"
aiogram-dialog = "^2.2.0"
pyyaml = "^6.0.2"
fluentogram = "^1.1.7"
alembic = "^1.13.3"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
psycopg = {extras = ["binary"], version = "^3.2.3"}


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
